通知设置大小字节颠倒

+++ TopStepComKit/Products/TopStepComKit.framework/Headers/TPSWebRequestModel.h
@@ -8,7 +8,6 @@
 #ifndef TPSWebRequestModel_h
 #define TPSWebRequestModel_h
 
-
 @interface TPSWebRequestModel : NSObject
 
 @property(nonatomic, strong) NSString* url;
@@ -17,7 +16,7 @@
  若存在云端下载的文件需要转发给手表，则要指定该文件在手表所存储的文件夹路径
  */
 @property(nonatomic, strong) NSString* remoteSavePath;
-@property(nonatomic, assign) int requestId;
+@property(nonatomic, assign) long requestId;
 @property(nonatomic, strong) NSString* appId;
 @property(nonatomic, strong) NSDictionary* params;
 @property(nonatomic, strong) NSArray* localPathArr;
